Support for struct promotion of ref types.
Getting Nth field of a ref type requires more work. Add support for this
in the main promotion loop, and stub out a few other places that get
confused trying to get the field name.

Be more aggressive promoting delegates. This is hacked in.
Delegate ref class is now promotable but doesn't get promoted because
we don't see the field uses (we only track these for address of local
structs). So hacked that part too.

That gets the delegate fields dependently promoted. Delegate as a whole
and all fields are address exposed. The "this" pointer is set to
&local + 8 to account for the preheader, and then that this pointer is
passed to Invoke. We later inline invoke and reveal that the delegate
is not actually exposed there, but too late for opt purposes.

Not sure yet if exposure totally blocks-per field modelling or only
selectively blocks it. Suspect the former but need to do some more digging.
Perhaps the delegate field values are visible at the call...